Adds affix dict support via aspell expand cmd
aspell supports dictionaries with affixes which determine the
grammatically valid expansions for each word. Affixes are separated from
the individual word via a forward slash. If such slash is found, the
word list has to be expanded. This is done via piping the word list
through aspell' 'expand' command with the language of the dictionary
given as parameter (if not set to "default").